dashboard changes, impliment admin and super admin feature

+++ webapp/app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Yajra\Datatables\Datatables;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Input;
+use Illuminate\Support\Facades\DB;
 use JsValidator;
 use App\User;
 
@@ -31,6 +32,7 @@ class UserController extends Controller
         'email' => 'required|email|unique:users,email',
         'password' => 'required|min:8|confirmed',
         'password_confirmation' => 'required',
+        'user_type' =>'required',
     );
 
     /**
@@ -51,19 +53,21 @@ class UserController extends Controller
     public function getData()
     {
         $users = User::get();
-
         
         return Datatables::of($users)
             ->addColumn('action', function ($users) {
-                $html =  '<a href="' . route('users.show', $users->id) . '" class="btn btn-xs btn-success" title="Show record"><i class="fa fa-eye"></i> Show</a>
-                <a href="' . route('users.edit', $users->id) . '" class="btn btn-xs btn-warning" title="Edit record"><i class="fa fa-edit"></i> Edit</a>';
+                $html =  '<a href="' . route('users.show', $users->id) . '" class="btn btn-xs btn-success" title="Show record"><i class="fa fa-eye"></i> Show</a>';
 
                 $currentLoginUserId = auth()->user()->id;
-                if($currentLoginUserId != $users->id){
-                    $html .= ' <button type="button" title="Delete record" class="btn btn-xs btn-danger btn-delete" data-placement="left" data-remote="' . route('users.destroy', $users->id) . '"><span class="fa fa-trash-o" aria-hidden="true"></span> Delete</button>';
-                }
-                else{
-                    $html .= ' <button type="button" disabled title="Delete record" class="btn btn-xs btn-danger btn-delete" data-placement="left" data-remote="' . route('users.destroy', $users->id) . '"><span class="fa fa-trash-o" aria-hidden="true"></span> Delete</button>';
+                $user_type = auth()->user()->user_type;
+                if($user_type == 'admin'){
+                    $html .= '<a href="' . route('users.edit', $users->id) . '" class="btn btn-xs btn-warning" title="Edit record"><i class="fa fa-edit"></i> Edit</a>';
+                    if($currentLoginUserId != $users->id){
+                        $html .= ' <button type="button" title="Delete record" class="btn btn-xs btn-danger btn-delete" data-placement="left" data-remote="' . route('users.destroy', $users->id) . '"><span class="fa fa-trash-o" aria-hidden="true"></span> Delete</button>';
+                    }
+                    else{
+                        $html .= ' <button type="button" disabled title="Delete record" class="btn btn-xs btn-danger btn-delete" data-placement="left" data-remote="' . route('users.destroy', $users->id) . '"><span class="fa fa-trash-o" aria-hidden="true"></span> Delete</button>';
+                    }
                 }
                 return $html;
             })
@@ -74,9 +78,16 @@ class UserController extends Controller
 
     public function create()
     {
-        $validator = JsValidator::make($this->validationRules, array(), array(), "form.create_user");
-
-        return view('users.create', compact('validator'));
+        $user_role = auth()->user()->user_type;
+        if($user_role == 'admin'){
+            $validator = JsValidator::make($this->validationRules, array(), array(), "form.create_user");
+            $user_type = User::select('user_type')->distinct()->get();
+            return view('users.create', compact('validator','user_type'));
+        }
+        else{
+            flash("You don't have this permission!")->warning()->important();
+            return redirect()->route('users.index');
+        }
     }
 
     /**
@@ -87,14 +98,20 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        //validate user data
-        $this->validate($request, $this->validationRules, array());
+        if(auth()->user()->user_type == 'admin'){
+            //validate user data
+            $this->validate($request, $this->validationRules, array());
 
-        //create new user
-        User::create($request->all());
+            //create new user
+            User::create($request->all());
 
-        flash('User created successfully!')->success()->important();
-        return redirect()->route('users.index');
+            flash('User created successfully!')->success()->important();
+            return redirect()->route('users.index');
+        }
+        else{
+            flash("You don't have this permission!")->warning()->important();
+            return redirect()->route('users.index');
+        }
     }
 
     /**
@@ -117,14 +134,20 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::findOrfail($id);
-        
-        //Js validation for front-end
-        $validationRules = array_except($this->validationRules, ['password', 'password_confirmation']);
-        $validationRules['email'] = $validationRules['email'] . "," . $id;
-        $validator = JsValidator::make($validationRules, array(), array(), "form.edit_user");
+        if(auth()->user()->user_type == 'admin'){
+            $user = User::findOrfail($id);
+            $user_type = User::select('user_type')->distinct()->get();
+            //Js validation for front-end
+            $validationRules = array_except($this->validationRules, ['password', 'password_confirmation']);
+            $validationRules['email'] = $validationRules['email'] . "," . $id;
+            $validator = JsValidator::make($validationRules, array(), array(), "form.edit_user");
 
-        return view('users.edit', compact('user', 'validator'));
+            return view('users.edit', compact('user', 'user_type' ,'validator'));
+        }
+        else{
+            flash("You don't have this permission!")->warning()->important();
+            return redirect()->route('users.index');
+        }
     }
 
     /**
@@ -136,25 +159,31 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $user = User::findOrFail($id);
-        
-        $input = $request->all(); 
-        if($request->password == ''){
-            $validationRules = array_except($this->validationRules, ['password', 'password_confirmation']);
-            unset($input['password']);
+        if(auth()->user()->user_type == 'admin'){
+            $user = User::findOrFail($id);
+            
+            $input = $request->all(); 
+            if($request->password == ''){
+                $validationRules = array_except($this->validationRules, ['password', 'password_confirmation']);
+                unset($input['password']);
+            }
+            else{
+                $validationRules = $this->validationRules;
+            }
+            $validationRules['email'] = $validationRules['email'] . "," . $id;
+            $validator = $this->validate($request, $validationRules, array());        
+
+            // update user
+            $user->fill($input);
+            $user->save();
+
+            flash('User updated successfully!')->success()->important();
+            return redirect()->route('users.index');
         }
         else{
-            $validationRules = $this->validationRules;
+            flash("You don't have this permission!")->warning()->important();
+            return redirect()->route('users.index');
         }
-        $validationRules['email'] = $validationRules['email'] . "," . $id;
-        $validator = $this->validate($request, $validationRules, array());        
-
-        // update user
-        $user->fill($input);
-        $user->save();
-
-        flash('User updated successfully!')->success()->important();
-        return redirect()->route('users.index');
     }
 
     /**
@@ -165,20 +194,26 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        // check current login user
-        $currentLoginUserId = auth()->user()->id;
-        if($currentLoginUserId == $id){
-            return response()->json([
-                'status' =>false,
-                'message' => "User can't delete itself !"
-            ]);            
-        }
+        if(auth()->user()->user_type == 'admin'){
+            // check current login user
+            $currentLoginUserId = auth()->user()->id;
+            if($currentLoginUserId == $id){
+                return response()->json([
+                    'status' =>false,
+                    'message' => "User can't delete itself !"
+                ]);            
+            }
 
-        //delete user from db
-        User::destroy($id);
-        return response()->json([
-            'status' =>true,
-            'message' => 'User deleted successfully!'
-        ]);
+            //delete user from db
+            User::destroy($id);
+            return response()->json([
+                'status' =>true,
+                'message' => 'User deleted successfully!'
+            ]);
+        }
+        else{
+            flash("You don't have this permission!")->warning()->important();
+            return redirect()->route('users.index');
+        }
     }
 }
